bug: Fixing NVMe get and set features in Windows
Fixing get and set NVMe features in Windows. The issue was how things have changed overtime in the way NVMe structures are passed down. Linux did not need the data pointer uniquely separate and could read it out of the NVMe command dwords, but the Windows code does not work this same way. So the pointer was not being setup properly resulting in data not coming back up the stack to the caller.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>