feat: More work towards realtek 9210
More work for supporting the realtek 9210 usb->nvme/sata adapter
There are multiple paths for detecting this and a possible warning will be emitted if a user has something that MIGHT be one of these, but was not otherwise detected due to differing MN and vendor ID.
This may require more tweaks going forward, but this is the best way I can figure we should support it for now.
Summary is if we see this adapter by USB vid/pid or known inquiry data we check the ATA identify response to see if DMA mode is supported (This should always be true on SATA drives) and if maxLBA is zero to determine if it is attached to an NVMe device and only issue it SCSI commands.

Unfortunately this adapter does not seem to support firmware updates like the original issue requested and realtek has been completely unresponsive to my emails for help on the capabilities of this adapter.

[Seagate/openSeaChest#109]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>