quick: Adding retry on synchronize cache 16
Adding a retry in case synchronize cache 16 to retry with 10B command instead.
Since Microsoft only implemented the 10B command for their SCSI to NVMe translator, I think it's pretty clear Microsoft is only issuing that command themselves. The 16B should translate the same, but was left out for some reason...
So now there is a retry in case any other device has the same issue.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>