feat: Adding initial PCI device hacks/workarounds
Adding some initial support for some reported issues on some SAS/SATA HBAs
One is the Redhat virtual HBA which limits SAT passthrough to A1h CDB, but it may have additional limitations that I do not know about at this time.
The other is a PMC controller that is blocking the SCSI write buffer command with deferred download mode set and fails the report supported operation codes command. It may have additional limitations, but I cannot test this right now.
I added a comment for Adaptec series-8 HBAs that ATA passthrough must use DMA mode as UDMA is blocked, however I didn't want to enable this right now without testing for other capabilities/limitations. Right now this is not an issue due to retries that are used in other parts of the library.

[Seagate/openSeaChest#47]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>