bug: Fixing odd rescan problem when SCSI and ATA reported data differs
Fixing a bug where if the maxLBA differs between read capacity and identify (which happens when you issue a SAT passthrough to change maxLBA), then rescanning later the device structure's maxLBA does not match the actual state of the drive.
This workaround will get an accurate maxLBA for the software to continue functioning as a user would expect.
Still working on finding a solution to update the SCSI translator's cached data and will make that change where appropriate when it is found.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>