bug: Changing infinite timeout on ATA set sector configuration to 15 minutes
Changing the timeout from infinite to 15 minutes.
In a case reported on Seagate/ToolBin#16 the drive was reset, but this status was not propagated back to the software (like it is supposed to be) so the drive was interrupted, but never had a chance to have a recovery command issued.
This change sets the timeout to 15 minutes. This should be more than enough for a device to complete the fast reformat and return since the average time it takes is 5 minutes already. At 15 minutes, this should handle outliers where the drive is just taking longer for unknown reasons.

[Seagate/ToolBin#16]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>