quick: Changing timeout on set sector configuration to 1 hour
Changing to a 1 hour timeout due to a few corner cases that could cause this to take longer. In all cases measured it completed long before an hour, but this is a time that is padded with extra room in case future products take more time, this will work around those possible time extensions for a while.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>