bug: Changing how the ATA workaround for single sector transfers works
Some ATA commands (Identify, DCO, Set Max, read/write buffer, etc) mark the sector count register as "N/A".
Often times this is interpreted as zero, so this workaround fixes a common SATL translation failure on these commands where it tries to figure out how large the data transfer is from the sector count register, but when set to zero it will crash the translator or fail the command.

The old workaround automatically applied sector count to 1 whenever a command had this set to zero, which caused some incorrect behavior in certain circumstances.

This new check improves when this workaround runs without needing a list of opcodes to check for.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>