bug: More changes to how IOCTL_STORAGE_REINITIALIZE_MEDIA is implemented
The IOCTL_STORAGE_REINITIALIZE_MEDIA behavior changes depending on a drive's capabilities and which version of the OS is currently running.

It's very confusing and not documented behavior.

These changes handle what seems to be the behavior from what I was able to test.
It is possible the version checks for the OS are not completely accurate, but since I don't have Windows 10 at each of these update levels to test individually or drives that meet criteria at each of these levels, it is very difficult to figure out exactly the behavior to expect since it has so many variables that are used internally but not documented.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>