bug: Changing the NVM format translation in Windows
Changing the NVM format translation in Windows because of the undocumented behavior that has been observed.
If you use the sanitize CDB, block erase = nvm sanitize block erase.
                             crypto erase = format + crypto erase.

I've also put in some support for calling the reinitialize media ioctl if all the conditions are right (drive support and Win 10 version) since that is clearly documented as starting in win10 1607, but the santize CDB translation is unclear.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>