quick: Clearing the NVMe completion data out before issueing reinit IOCTL
Clearing out the completion data before issuing the reinitialize media IOCTL to make sure it is free of any data since this IOCTL will not tell us the completion registers.
It may be possible to "guess" them from status codes from Windows, but until we can repeat an error with a bus trace, this will be left empty.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>