quick: Adding another case to handle
While reviewing all the changes for when to issue this IOCTL, I missed when it is before Win10 21H1 and after 1903 which the driver can know how to issue sanitize crypto erase.
So I have added that case in here to support that case as well.
If we ever get enough information to further refine this check based on Windows 10/11 subversions, we should implement it based on whatever we can find out.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>