quick: Cleaning up some changes for HpSAMD.sys
Cleaning up the changes I had tested to get this driver to respond since it seems that this driver will not respond to passthrough requests at all.
It does support the miniport registry key, but changing it to CSMI=Full; does not seem to make a difference. In fact, setting Full seems to disable all CSMI functionality for some unknown reason.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>