bug: Removing NVMe format translation
Removing the NVMe format unit translation code because with Microsoft's recent MSDN updates, it seems that this is not actually supported. This seems to match my test results as well.
It is possible that this is supported in an older Windows 10 update, but that information is not available.
Everything we've tested recently only seems to support sanitize crypto erase.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>