feat: Getting CISS support working in Linux
Getting the CISS support working in Linux with scan and the "normal" get_Device path.
Been testing this on my end and so far everything is working as expected.

There is still some cleanup/tweaks to do to further refine this since I am seeing the drives show up duplicated between sg0 and sg1 handles, so looking into a solution for that still.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>