feat: Adding a few optimizations to detect when to set some SCSI hacks
Adding in detecting when mode pages, log pages, and report supported opcodes commands are not supported.
This currently relies on sense key specific information to come back and specify an invalid field to be correctly handled, but will help in certain cases of unsupported operation codes as well.
These do not block issuing these commands, but setup the necessary information to help upper layers determine whether to keep issuing these commands or not.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>