doc: Updating opensea-transport license for the files added for CISS support.
Adding the FreeBSD and Illumos licenses since opensea-transport is using files from their kernels to enable support for issuing commands to drives behind these RAIDs.
Note: This is not currently enabled for Illumos or FreeBSD builds at this time. It is planned to be turned on in the future when we can test it on these OSs.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>