doc: Adding comment about automatic scsi hack for reason this is ok
Adding a comment that describes why it is considered "OK" to disable reading mode pages  in the hacks list when the mode sense 6 fails.
This does not block the ability to issue the command in opensea-transport, but it is useful to know and choose whether to try again or not.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>