feat: Autodetecting when SAT VPD page is not supported
Auto detecting when the SAT VPD page is not supported because the device did not report it in the list of supported VPD pages.
This was otherwise always being attempted due to how this code looks for SAT translators being present.
This allows skipping a command and going straight towards a passthrough identify test instead which is what most USB devices require anyways.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>