feat: Improving autodetection of SCSI hacks for LP, MP, VPD
Improving autodetection of support for log pages, mode pages, and VPD pages.
In order to get this to work correctly, the code has to track if there have been any successful requests in the past already or not.
Some USB devices do support these, but return odd error codes, so there also needs to be a limit of how many attempts are required before deciding something is not supported. These limits are currently set to 5 to make sure there are enough possible attempts to commonly supported pages before giving up.
This may need further tweaking in the future.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>