bug: Some sntl code was using ATA identify data instead of nvme
Fixed the SNTL code that was using ATA identify data to detect capabilities and features instead of NVMe identify data.
This may have lead to incorrect translations of the data-set management command and sanitize commands.
It is unlikely many users saw these errors as most of the time the upper layers of this code will natively use NVMe commands and not SCSI CDBs to interact with NVMe devices.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>