bug: Fixing SAT for PIO-in and warning about incomplete RTFRs
SAT was incorrectly failing some identify commands (or other PIO-in) because the RTFRs were zero.
This is not the correct way to handle PIO-in or FPDMA. For these protocols, a successful command may not return any information from some SATLs because the command was successful. If these commands fail, then they will return a failure code.

Also found that the warning about incomplete RTFRs was not working so I also fixed that error again.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>