quick: Disable SAT check condition for pio-in and fpdma
Disabling setting the check condition bit for pio-in and fpdma commands when building a CDB and the SATL supports setting it on all commands.
These protocols will not return completion as expected and SAT even specifies this. Some translators will just return zeroes on these protocols as well.
The only time the check condition can be set for these is if the force check condition bit is set.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>