feat: Detecting SATA sanitize definitive ending pattern support
Adding in detection for SATA's sanitize overwrite definitive ending pattern support bit from the ID data log.
When this is set, behavior matches definitions in SCSI and NVMe sanitize standards for the final pattern regardless of passes and inversion of the pattern.
When this is not set, the device may or may not set the
This is being set by default when issuing the Sanitize command to make sure the behavior is as expected compared to the other standards.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>