feat: Converting int returns to eReturnValues
Converting the int return types to eReturnValues since this is what these functions were using to return a status anyways.
This was a massive change throughout all libraries to make this change, but it is more correct and doesn't rely on a silent conversion which throws off C++ code that uses these functions.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>