bug: Fixing macro used to get logical sector size on ATA drives
This data is supposed to be read from 2 words as a dword, but we were using the bytes macro which could have been causing the logical sector size to be interpreted incorrectly.
I am not aware of a real case where this was happening at this time, but fixing this before it becomes a problem.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>