make: Moving NVMe detection from opensea-common to opensea-transport where it belongs
Moving the detection of NVMe passthrough support to opensea-transport from opensea-operations in meson.
This is leftover from how this used to be detected long ago and how the ifdef was used in the code.

Today this ifdef is only used in opensea-transport and does not get used to wrap other NVMe functionality in other libraries anymore since this only controls if there is an NVMe passthrough available in a given OS. The upper-layers still define NVMe functions even if a passthrough is not available to be able to support USB devices that have NVMe drives on them.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>