feat: Adding support for VMWare NVM IO.
Added in code to support VMWare NVM IO in addition to admin passthrough.
This seems to be rejected with an error for "Not supported" and I do not know why, but it is what it is.
I also cleaned up all warnings from compilation and removed unused/dead code.

There was also a bug in the handle allocation that was also resolved to make sure we stay within memory bounds.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>