bug: Removing special case cast in VMWare for NVMe get log page
This special case to cast to uint32 is no longer needed and does not work at this level. The appropriate code to handle setting the pointer is now in the vm_helper.c file.
By moving that cast down to there with some other changes, this command is now working as expected.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>