bug: Fixing drive_type not being set correctly while parsing sysfs
Fixing a bug that can occur for some devices where the drive_type does not end up being set at all and is left as "UNKNOWN".
Many of the upper layers are expecting ATA, SCSI, or NVMe and if left as unknown it requires forcefully setting the drive type.
By setting it at this low-level it corrects this behavior without other workarounds by users

[Seagate/openSeaChest#155]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>