feat: Adding inline functions to declare ATA commands and reduce boilerplate
Many ATA functions get built with almost identical fields in place in the same boiler plate code. I have created inline functions for the most common parts in order to improve consistency, reduce boilerplate code, and prevent bugs if a field was setup incorrectly for common formats. While I do not know of a current case of a bug from this, it has been observed in the past and this will keep things easier to maintain.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>