feat: Adding new USB device to list and auto-detection for future versions
There have been a few devices that show up with some common information so now if a new one is released a software update will not be necessary to support it. Adding more PIDs is still the best method but this includes a workaround to detect these seemingly common devices.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>