feat: Adding new function to make better retries for initial ATA passthrough identify
Adding a function to handle which things to attempt for retries as the code tries it's first identify command to determine if passthrough is available or not.
This allows for more retries than previous code to handle a few other things to try as well as handling mis behaving USB devices that sometimes return the incorrect sense code.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>