quick: Using macro MSFT_BOOL_<TRUE|FALSE>
Using the MSFT_BOOL_TRUE and MSFT_BOOL_FALSE macros to ensure that the comparisons are correct.
Since BOOL in Windows is a typedef to int, this means FALSE is 0 and TRUE is 1, however if an API call were to return anyother non-zero value it may not be interpretted quite as expected. These macros ensure correct interpretation by comparing == FALSE or != FALSE rather than direct comparisons to TRUE

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>