quick: Fixing warning about nonnull compare
Fixing nonnull compare warning for AIX, FreeBSD, Illumos, Solarir, and ESXI

The issue is since the get_Device_List requires nonnull parameters we get a warning when checking for null.
So it now explicitly checks for null and disables the nonnull compare warning in this section since we still want to make sure a NULL pointer is not passed in case the warning is ignored.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>