bug: Fixing how the adapter ID is read in Linux
Fixing how the adapter IDs are read in linux.
The getline to these files has some extra control characters at the end, so added a function to trim those off after they are read into memory and before we try parsing them.

In process of debugging reading the driver information so some extra prints from that are still in place.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>