bug: Fixing reading driver info in Linux
Fixing a bug reading the driver version info under linux.
The file is marked as a size of 4096, but only has a few characters in it.
Rather than let this trigger a failure to read all the data, I switched to getline to read the data, then parse it instead.
This seems to work and is simpler overall and fixes the parsing.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>