feat: Converting inline scsi sense check functions to regular functions
Converting some functions that checked SCSI sense data from inline functions to regular functions. This was done in order to fix compatibility between static libraries and DLLs in Windows.
I also simplified the code with a single place to do the check and made these helper functions call that one single function for the checking to reduce duplication and improve ease of use.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>