feat: Adding initial support for Sunplus USB ATA Passthrough support
Adding initial support for Sunplus's ATA Passthrough capabilities.
I tested this on one adapter I had and it appears to work from the testing that I did.
This is limited to PIO commands only but that covers a lot of what we need.
I did test PIO reads and those also worked just fine.

[Seagate/openSeaChest#197]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>