bug: Feature/nvme win11 fwdl updates (#24)
Windows 11 introduced new flags in the last two updates to the firmware update API.
This adds detection of the running version of the OS to choose the best supported method to activate the firmware based on Windows support and drive support.
Win10 and Win11 both support the replace and activate on reset option, Win11's new API is just more verbose on which method to choose.
This also supports both when the FWDL API is available and when it is not available to choose the best activation method (Windows10/11 vs some other NVMe driver or USB adapter).

[Fixes Seagate/ToolBin#39, Fixes Seagate/openSeaChest#198]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>
# Conflicts:
#	include/common_public.h
#	include/nvme_helper.h
#	include/win_helper.h
#	src/cmds.c
#	src/win_helper.c