bug: Fixing scan not showing any output in specific situations
The way in which scanning the system determined when to show or not show drives had a bug where if the number of failed to scan and number of successfully scanned drives were the same, nothing would show up.
While addressing this I added additional checks for busy so that if all devices are busy, that can be propagated upwards too.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>