bug: Fixing scan not showing any output in specific situations
The way in which scanning the system determined when to show or not show drives had a bug where if the number of failed to scan and number of successfully scanned drives were the same, nothing would show up.
While addressing this I added additional checks for busy so that if all devices are busy, that can be propagated upwards too.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

# Conflicts:
#	src/aix_helper.c
#	src/ata_cmds.c
#	src/cam_helper.c
#	src/common_public.c
#	src/netbsd_openbsd_helper.c
#	src/sg_helper.c
#	src/uscsi_helper.c
#	src/win_helper.c