bug: Fixing a strange bug when issuing a CDB and copying sense data to tDevice structure
Fixing a weird error that was found when calling the send_CDB function from python.
An error would sometimes occur during the copy of sense data back to tDevice->drive_info.lastCommandSenseData

The issue would not always occur. It also seemed to depend on calling speed in some cases too, which made it even odder.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>