bug: Moving copy to lastCommandSenseData to private send CDB function
Moving the copy of the last command sense data to the private send CDB function since this is the "root" function responsible for issuing CDBs and should also be the one doing this copy.

Also added a memset to sense data before issuing a command to ensure it is cleared out to minimize the chance of a command reporting incorrectly.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>