merge: Feature/cleanup (#26)
* make: Updating clang-tidy configuration

updating the clang-tidy config file for new features.
Some are things we should come back to enable, and others are disabling things we are not ready to support just yet.
We are disabling bugprone assignment in condition since we use that in a few places heavily and will require a big refactor.
We are disabling a modernization for macro to enum since that wants to use annonymous enums which is a C11 standard feature and we are not quite ready to adopt C11.
Finally disabling a include since we do handle some system oddness with our own files we include for portability and this may not otherwise solve the issue it is warning about.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* feat: Set features refactor, default timeout macro, other clang-tidy fixes

Refactoring ATA set features to have helper functions for each defined ATA set features command from the standards and specification.
Defining a macro for a default timeout to increase readability.
Other clang-tidy fixes to resolve other warnings and reduce complexity of some functions.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* bug: Fixing linking error in VS2022

Not sure why gcc and clang worked, but VS2022 would not link this without fixing the function names.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* bug: Fixing CSMI error by adding explicit cast from __u32 to uint32_t

Generic selection version of M_Word0 did not like the conversion without a cast and this resolved it.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* feat: Converting inline scsi sense check functions to regular functions

Converting some functions that checked SCSI sense data from inline functions to regular functions. This was done in order to fix compatibility between static libraries and DLLs in Windows.
I also simplified the code with a single place to do the check and made these helper functions call that one single function for the checking to reduce duplication and improve ease of use.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* make: Upping strict-flex-arrays to 3 and reworking supported formats API for this as well

Upping the strict-flex-arrays to 3 to make sure that all possible array overruns (reads or writes) are caught to enhance security even further.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* bug: Fixing incorrect protocol on read stream ext pio and incorrect opcode on trusted receive pio

Fixing incorrect protocol for read stream ext in PIO mode and fixing the provided opcode for trusted receive pio.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* quick: clang-format changes

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* feat: Reworking the ata command creation helper functions for better type checking

Reworking how the helper functions to create an ATA command work to use enumeration types for better type checking to avoid errors

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* feat: switching printf using no formatting to print_str call added to io_utils

Switching to print_str from printf where printf was extra overhead that was not needed.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* feat: Small cleanup to how to issue sanitize overwrite

Cleaning up the code to issue sanitize overwrite slightly by using a common variable for ATA and NVMe patterns and reusing that when the pattern is null for SCSI devices.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* quick: Adding io_utils.h for of_nvme_helper

Adding io_utils.h to resolve a warning

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* quick: Adding some more definitions for self test codes, log lengths, scsi defect descriptors

Adding enums with more values so that they can be used for more readable code instead of hard-coded values

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* bug: Fixing missing function print_str

adding io_utils.h for print_str symbol

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* bug: Removing UINT16_C around DEFAULT_COMMAND_TIMEOUT macro to fix uscsi_helper

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

---------

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>