feat: Initial checking of AACRAID support for older adaptec controllers
Initial changes to check in to support AACRAID driver. This is for Adaptec cards using this driver in RAID mode.
It was written based off of what was in the linux kernel driver, freebsd kernel driver, and illumos kernel driver to figure out how to get it to work.

Note that at this time it is not fully functional, this is just the base code to work from and will require further modifications

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>