feat: FreeBSD NVMe CAM updates and Legacy ATA support (#29)
* feat: Adding some legacy ATA Passthrough support

Adding some legacy ATA Passthrough support for old kernels.
This may seem odd since this was last supported in FreeBSD 8, but we get a lot of odd requests, so this support ensures we can handle requests for old systems.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* feat: Adding cam nvme support to use for /dev/nda devices

CAM added NVMe support so this pulls in support for this interface as well.
Currently this change makes CAM the default choice for all device types which keeps this code fairly simple.
The NVMe IOCTL paths are still supported right now, however they are not default. We may want to add some kind of flag to switch between the two at runtime from openSeaChest like -s -F preferNVMeIOCTL.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* quick: Adding one more check for NVMe before including nvme_all

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* bug: Fixing compilation issues on FreeBSD 13.5

Fixing incorrect legacy ATA structure access and some NVMe CAM issues that were found when attempting to compile.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* quick: Fixing DISABLE_NVME_PASSTHROUGH flag check

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* quick: Fix pointer access to the device name and filedes

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* quick: Fixing a few warnings and making the legacy ATA PT a static function

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* bug: Making is_NVMe_Handle take const char* and fixing missing semicolon

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* bug: Fixing compilation on FreeBSD 13.5

Fixing bugs trying to compile on FreeBSD 13.5
Missed setting a defined value to a number and there are newer NVMe fabrics work that is not present in older freebsd.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* bug: Fixing invalid interface errors in FreeBSD

Fixing an invalid interface error that can show for USB devices and firewire (1394) devices.
This was due to not handling those interfaces when choosing how to issue a SCSI passthrough command.
Instead of if-else, this now uses a switch-case to make sure we have them handled.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* quick: Fixing this stupid file with no changes that git keeps complaining about

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* quick: Fixing some errors from merging in new process for common open handle method

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* bug: Fixing dragonfly bsd build after merge

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* bug: One more change for NVMe handle detection to be compatible with dragonfly bsd

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* quick: Fix unused function warning in Dragonfly BSD

Fixing unused function warning for DragonFly BSD and the legacy ATA PT support.
The function was not being called for Dragonfly builds but was called for FreeBSD builds.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* quick: Fixing length passed to getfsstat and getvfsstat in netbsd and openbsd

The man page shows a size_t so conversion to an int is not what we want to do.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

---------

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>